Change LSN fields from long to decimal
SQL Server LSN columns (last_hardened_lsn, last_commit_lsn) are
numeric(25,0) which can exceed Int64.MaxValue (~19 digits). Changed
all LSN properties in DatabaseReplicaState, LsnComparisonResult, and
HealthLevelExtensions from long to decimal. Use reader.GetDecimal()
for LSN columns in both AgMonitorService and DagMonitorService.

+++ src/SqlAgMonitor.Core/Models/HealthLevel.cs
@@ -14,7 +14,7 @@ public enum HealthLevel
 
 public static class HealthLevelExtensions
 {
-    public static HealthLevel FromLsnDifference(long lsnDifference, bool isDisconnected = false)
+    public static HealthLevel FromLsnDifference(decimal lsnDifference, bool isDisconnected = false)
     {
         if (isDisconnected) return HealthLevel.DangerZone;
